iium-vectordb/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Phase 1: Ingestion ────────────────────────────────────────────────────
google-api-python-client>=2.100.0    # Google Drive API
google-auth>=2.23.0                  # Service account auth
gdown>=5.1.0                         # Public-folder fallback
pdfplumber>=0.10.0                   # PDF parsing
python-docx>=1.1.0                   # DOCX parsing

# ── Phase 2: Embeddings ───────────────────────────────────────────────────
openai>=1.30.0                       # text-embedding-3-small

# ── Phase 3: Vector DB ────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Phase 4: API ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Phase 6: Stress test ──────────────────────────────────────────────────
httpx>=0.27.0
